+++ packages/zeebe-element-templates-json-schema/CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to [@camunda/zeebe-element-templates-json-schema](https://gi
 
 ___Note:__ Yet to be released changes appear here._
 
+* `FEAT`: allow `zeebe:agentDefinition` binding on `bpmn:ServiceTask` and `bpmn:AdHocSubProcess` ([#265](https://github.com/camunda/element-templates-json-schema/pull/265))
+
 ## 0.45.0
 
 * `FEAT`: allow `businessId` as `zeebe:calledElement` binding property ([#263](https://github.com/camunda/element-templates-json-schema/pull/263))